backend-django/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# File Upload & Storage
Pillow==10.1.0
django-storages==1.14.2

# API Documentation
drf-spectacular==0.27.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
factory-boy==3.3.0
pytest-django==4.7.0

# AI Integration
openai==1.3.7

# Utilities
celery==5.3.4
redis==5.0.1
requests==2.31.0

# Real-time and Data Processing
channels==4.0.0
channels-redis==4.1.0
websockets==12.0

# Data Visualization Support
pandas==2.1.4
numpy==1.26.2

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0